+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Make identity proof submission work with new API.
+- Send chain ID to server when updating subscription parameters.
 
 ## [1.28.0] - 2023-06-26